+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed `EUiDataGrid`'s cell popover overlapping with modals and flyouts ([#5461](https://github.com/elastic/eui/pull/5461))
 - Fixed an accessibility issue where `EuiDatePicker` time options did not have unique IDs ([#5466](https://github.com/elastic/eui/pull/5466))
+- Fixed global and reset styles when using the legacy theme ([#5473](https://github.com/elastic/eui/pull/5473))
 
 ## [`43.1.0`](https://github.com/elastic/eui/tree/v43.1.0)